Nest Http2's Stream, and stop the tests importing Http2 wholesale
`Http2Stream` becomes `Http2.Stream` after all. The ambiguity that
prompted reverting it was not really the nesting's fault: cordillera's
tests did `import Http2.*`, so every member of the object landed in a
scope that already had `import soundness.*`, and any name shared with
the umbrella — `Stream` here — became ambiguous. A wildcard over a
namespace object is what makes nesting fragile, so the test imports
only the nine members it uses.

The library's own files keep their `import Http2.*`: importing the
members of an object from inside the file that defines it is a
different thing from a consumer doing so.
