--- lib/telekinesis/src/http2/cordillera.Http2Stream.scala
@@ -1,82 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.64.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package cordillera
-
-import scala.caps
-import proscenium.compat.*
-import scala.collection.mutable as scm
-import anticipation.{Data as Bytes, *}
-import coaxial.*
-import contingency.*
-import fulminate.*
-import gossamer.*
-import parasite.*
-import prepositional.*
-import rudiments.*
-import spectacular.*
-import telekinesis.*
-import vacuous.*
-import Http2.Error.Reason
-import anticipation.*
-import java.util.concurrent.atomic as juca
-import scala.collection.concurrent as scc
-import hieroglyph.*, charEncoders.asciiEncoder
-import proscenium.*
-import turbulence.*
-import zephyrine.*
-import Http2.*
-
-class Http2Stream(val id: Int):
-  val headers: Promise[List[HpackEntry]] = Promise()
-  val trailers: Promise[List[HpackEntry]] = Promise()
-  val body: Relay[Bytes] = Relay()
-  // Untracked: written only by the connection's single reader daemon.
-  @caps.unsafe.untrackedCaptures
-  private var headersSeen: Boolean = false
-
-  // Record an incoming HEADERS block: the first becomes the response headers, a
-  // subsequent one (always end-stream) becomes the trailers.
-  def acceptHeaders(block: List[HpackEntry]): Unit =
-    if !headersSeen then
-      headersSeen = true
-      headers.offer(block)
-    else
-      trailers.offer(block)
-
-  def acceptData(data: Bytes): Unit = body.put(data)
-
-  // Close the receive side; resolve any unfulfilled promises so awaiters don't hang.
-  def end(): Unit =
-    if !headers.ready then headers.offer(Nil)
-    if !trailers.ready then trailers.offer(Nil)
-    body.stop()
